requirements.txt
arrow==0.15.5
asgiref==3.3.0
binaryornot==0.4.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cookiecutter==1.7.0
Django==2.1.5
django-tinymce4-lite==1.8.0
future==0.18.2
idna==2.8
Jinja2==2.10.3
jinja2-time==0.2.0
jsmin==2.2.2
MarkupSafe==1.1.1
poyo==0.5.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
whichcraft==0.6.1
